+++ python/neuromation/cli/job.py
@@ -97,6 +97,13 @@ def job() -> None:
     help="Request extended '/dev/shm' space",
 )
 @click.option("--http", type=int, help="Enable HTTP port forwarding to container")
+@click.option(
+    "--http-auth/--no-http-auth",
+    is_flag=True,
+    help="Enable HTTP authentication for forwarded HTTP port",
+    default=True,
+    show_default=True,
+)
 @click.option("--ssh", type=int, help="Enable SSH port forwarding to container")
 @click.option(
     "--preemptible/--non-preemptible",
@@ -106,7 +113,19 @@ def job() -> None:
     show_default=True,
 )
 @click.option(
-    "-d", "--description", metavar="DESC", help="Add optional description to the job"
+    "-n",
+    "--name",
+    metavar="NAME",
+    type=str,
+    help="Optional job name",
+    default=None,
+    show_default=True,
+)
+@click.option(
+    "-d",
+    "--description",
+    metavar="DESC",
+    help="Optional job description in free format",
 )
 @click.option(
     "-q", "--quiet", is_flag=True, help="Run command in quiet mode (print only job id)"
@@ -148,12 +167,14 @@ async def submit(
     memory: str,
     extshm: bool,
     http: int,
+    http_auth: bool,
     ssh: int,
     cmd: Sequence[str],
     volume: Sequence[str],
     env: Sequence[str],
     env_file: str,
     preemptible: bool,
+    name: str,
     description: str,
     quiet: bool,
     wait_start: bool,
@@ -213,7 +234,7 @@ async def submit(
         log.debug(f"IMAGE: {parsed_image}")
     image_obj = Image(image=parsed_image.as_repo_str(), command=cmd)
 
-    network = NetworkPortForwarding.from_cli(http, ssh)
+    network = NetworkPortForwarding.from_cli(http, ssh, http_auth)
     resources = Resources.create(cpu, gpu, gpu_model, memory, extshm)
     volumes = Volume.from_cli_list(username, volume)
     if volumes and not quiet:
@@ -229,6 +250,7 @@ async def submit(
             network=network,
             volumes=volumes,
             is_preemptible=preemptible,
+            name=name,
             description=description,
             env=env_dict,
         )
@@ -272,7 +294,8 @@ async def exec(
 
 @command(context_settings=dict(ignore_unknown_options=True))
 @click.argument("id")
-@click.argument("local_port")
+@click.argument("local_port", type=int)
+@click.argument("remote_port", type=int)
 @click.option(
     "--no-key-check",
     is_flag=True,
@@ -280,14 +303,16 @@ async def exec(
 )
 @async_cmd
 async def port_forward(
-    cfg: Config, id: str, no_key_check: bool, local_port: int
+    cfg: Config, id: str, no_key_check: bool, local_port: int, remote_port: int
 ) -> None:
     """
     Forward a port of a running job exposed with -ssh option
     to a local port.
     """
     async with cfg.make_client() as client:
-        retcode = await client.jobs.port_forward(id, no_key_check, local_port, 22)
+        retcode = await client.jobs.port_forward(
+            id, no_key_check, local_port, remote_port
+        )
     sys.exit(retcode)