+++ python/tests/e2e/conftest.py
@@ -12,7 +12,7 @@ from hashlib import sha1
 from os.path import join
 from pathlib import Path
 from time import sleep, time
-from typing import List
+from typing import List, Optional
 from uuid import uuid4 as uuid
 
 import aiohttp
@@ -29,7 +29,13 @@ from neuromation.client import (
     ResourceNotFound,
 )
 from neuromation.utils import run
-from tests.e2e.utils import FILE_SIZE_B, RC_TEXT
+from tests.e2e.utils import (
+    FILE_SIZE_B,
+    JOB_TINY_CONTAINER_PARAMS,
+    NGINX_IMAGE_NAME,
+    RC_TEXT,
+    JobWaitStateStopReached,
+)
 
 
 JOB_TIMEOUT = 60 * 5
@@ -37,7 +43,8 @@ JOB_WAIT_SLEEP_SECONDS = 2
 JOB_OUTPUT_TIMEOUT = 60 * 5
 JOB_OUTPUT_SLEEP_SECONDS = 2
 STORAGE_MAX_WAIT = 60
-
+CLI_MAX_WAIT = 180
+NETWORK_TIMEOUT = 60.0 * 3
 
 DUMMY_PROGRESS = ProgressBase.create_progress(False)
 
@@ -48,6 +55,15 @@ job_id_pattern = re.compile(
     r"(job-[0-9a-f]{8}-[0-9a-f]{4}-[4][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12})",
     re.IGNORECASE,
 )
+
+
+@pytest.fixture
+def loop():
+    loop = asyncio.new_event_loop()
+    asyncio.set_event_loop(loop)
+    yield loop
+    loop.run_until_complete(loop.shutdown_asyncgens())
+    loop.close()
 
 
 class TestRetriesExceeded(Exception):
@@ -302,7 +318,9 @@ class Helper:
             job = await client.jobs.status(job_id)
             while job.status == wait_state and (int(time() - start_time) < JOB_TIMEOUT):
                 if stop_state == job.status:
-                    raise AssertionError(f"failed running job {job_id}: {stop_state}")
+                    raise JobWaitStateStopReached(
+                        f"failed running job {job_id}: {stop_state}"
+                    )
                 await asyncio.sleep(JOB_WAIT_SLEEP_SECONDS)
                 job = await client.jobs.status(job_id)
 
@@ -313,7 +331,9 @@ class Helper:
             job = await client.jobs.status(job_id)
             while target_state != job.status:
                 if stop_state == job.status:
-                    raise AssertionError(f"failed running job {job_id}: '{stop_state}'")
+                    raise JobWaitStateStopReached(
+                        f"failed running job {job_id}: '{stop_state}'"
+                    )
                 if int(time() - start_time) > JOB_TIMEOUT:
                     raise AssertionError(
                         f"timeout exceeded, last output: '{job.status}'"
@@ -373,8 +393,9 @@ class Helper:
 
         log.info("Run 'neuro %s'", " ".join(arguments))
 
+        t0 = time()
         delay = 0.5
-        for i in range(5):
+        while time() - t0 < CLI_MAX_WAIT:  # wait up to 3 min
             pre_out, pre_err = self._capfd.readouterr()
             pre_out_size = len(pre_out)
             pre_err_size = len(pre_err)
@@ -386,6 +407,7 @@ class Helper:
                             "--show-traceback",
                             "--disable-pypi-version-check",
                             "--color=no",
+                            f"--network-timeout={self.config.network_timeout}",
                         ]
                         + arguments
                     )
@@ -488,6 +510,8 @@ def config(tmp_path, monkeypatch):
             config = rc.ConfigFactory.load()
     else:
         config = rc.ConfigFactory.load()
+
+    config.network_timeout = NETWORK_TIMEOUT
     yield config
 
 
@@ -527,6 +551,8 @@ def config_alt(tmp_path, monkeypatch):
             config = rc.ConfigFactory.load()
     else:
         pytest.skip("CLIENT_TEST_E2E_USER_NAME_ALT variable is not set")
+
+    config.network_timeout = NETWORK_TIMEOUT
     yield config
 
 
@@ -573,3 +599,40 @@ def nested_data(static_path):
     nested_dir.mkdir(parents=True, exist_ok=True)
     generated_file, hash = generate_random_file(nested_dir, FILE_SIZE_B)
     return generated_file, hash, str(root_dir)
+
+
+@pytest.fixture
+def secret_job(helper):
+    def go(http_port: bool, http_auth: bool = False, description: Optional[str] = None):
+        secret = str(uuid())
+        # Run http job
+        command = (
+            f"bash -c \"echo -n '{secret}' > /usr/share/nginx/html/secret.txt; "
+            f"timeout 15m /usr/sbin/nginx -g 'daemon off;'\""
+        )
+        args = []
+        if http_port:
+            args += ["--http", "80"]
+            if http_auth:
+                args += ["--http-auth"]
+            else:
+                args += ["--no-http-auth"]
+        if not description:
+            description = "nginx with secret file"
+            if http_port:
+                description += " and forwarded http port"
+                if http_auth:
+                    description += " with authentication"
+        args += ["-d", description]
+        http_job_id = helper.run_job_and_wait_state(
+            NGINX_IMAGE_NAME, command, JOB_TINY_CONTAINER_PARAMS + args
+        )
+        status: JobDescription = helper.job_info(http_job_id)
+        return {
+            "id": http_job_id,
+            "secret": secret,
+            "ingress_url": status.http_url,
+            "internal_hostname": status.internal_hostname,
+        }
+
+    return go